third_party/blink/web_tests/external/wpt/editing/run/delete_5001-6000-expected.txt

This is a testharness.js-based test.
Found 13 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<div>foo[bar</div><p>baz]quz" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["delete",""]] "<p>foo[bar<p style=color:blue>baz]quz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["delete",""]] "<p>foo[bar<p style=color:blue>baz]quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<font color=\\"#0000ff\\">quz</font></p>" but got "<p>foo<span style=\\"color:rgb(0, 0, 255)\\">quz</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["delete",""]] "<p>foo[bar<p style=color:blue>baz]quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<font color=\\"#0000ff\\">quz</font></p>" but got "<p>foo<span style=\\"color:rgb(0, 0, 255)\\">quz</span></p>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<div><p>foo<p>[bar<p>baz]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p><p><br></p></div>" but got "<p>foo</p><p><br></p>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<div><p>foo<p>[bar<p>baz]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p><p><br></p></div>" but got "<p>foo</p><p><br></p>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<p>foo[</p><p>]bar<br>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar<br>baz</p>" but got "<p>foobar</p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<p>foo[</p><p>]bar<br>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar<br>baz</p>" but got "<p>foobar</p><p>baz</p>"
[FAIL] [["delete",""]] "<table><tbody><tr><th>fo[o<th>bar<th>b]az<tr><td>quz<td>qoz<td>qiz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><th>fo</th><th><br></th><th>az</th></tr><tr><td>quz</td><td>qoz</td><td>qiz</td></tr></tbody></table>" but got "<table><tbody><tr><th>fo</th><th></th><th>az</th></tr><tr><td>quz</td><td>qoz</td><td>qiz</td></tr></tbody></table>"
[FAIL] [["delete",""]] "<table><tbody><tr><th>[foo<th>bar<th>baz]<tr><td>quz<td>qoz<td>qiz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><th><br></th><th><br></th><th><br></th></tr><tr><td>quz</td><td>qoz</td><td>qiz</td></tr></tbody></table>" but got "<table><tbody><tr><th><br></th><th></th><th><br></th></tr><tr><td>quz</td><td>qoz</td><td>qiz</td></tr></tbody></table>"
[FAIL] [["delete",""]] "<table><tbody><tr><th>[foo<th>bar<th>baz<tr><td>quz<td>qoz<td>qiz]</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><th><br></th><th><br></th><th><br></th></tr><tr><td><br></td><td><br></td><td><br></td></tr></tbody></table>" but got "<br>"
[FAIL] [["delete",""]] "<table><tbody><tr><td>foo<td>ba[r<tr><td>baz<td>quz<tr><td>q]oz<td>qiz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td>ba</td></tr><tr><td><br></td><td><br></td></tr><tr><td>oz</td><td>qiz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td>ba</td></tr><tr><td>oz</td><td>qiz</td></tr></tbody></table>"
[FAIL] [["delete",""]] "<p>foo<table><tr><td>ba[r</table><p>b]az" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><table><tbody><tr><td>ba</td></tr></tbody></table><p>az</p>" but got "<p>foo</p><table><tbody><tr><td>baaz</td></tr></tbody></table>"
Harness: the test ran to completion.
